refactor: modularize CLI into separate command files
- Split main.rs (913 lines) into modular structure
- Create commands/ directory with individual command modules
- Extract shared utilities to util.rs
- Update examples to include heb (Hebrew) and vec (Venetian)
