+++ crates/unimorph-cli/src/commands/mod.rs
@@ -1,0 +1,19 @@
+//! CLI command implementations.
+
+mod analyze;
+mod delete;
+mod download;
+mod export;
+mod inflect;
+mod list;
+mod search;
+mod stats;
+
+pub use analyze::cmd_analyze;
+pub use delete::cmd_delete;
+pub use download::cmd_download;
+pub use export::{ExportFormat, cmd_export};
+pub use inflect::cmd_inflect;
+pub use list::cmd_list;
+pub use search::cmd_search;
+pub use stats::cmd_stats;
